MNT #66 refactor
